move purge uploads directory to its own file
